// src/components/exhibition/dto/booth-organization-product.dto.ts
/**
 * COMPAON API
 * API definition for COMPAON
 *
 * The version of the OpenAPI document: 1.0.0
 * Contact: compaon@dalavina.com
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { BoothOrganizationTemplatePosition } from './booth-organization-template-position.dto';

export interface BoothOrganizationProduct {
    readonly id: number;
    name: string;
    selected_media_id?: number;
    readonly image_id: number;
    media_data?: string;
    price: number;
    purchase_link: string;
    description: string;
    booth_organization_template_position_id: number;
    booth_organization_template_position?: BoothOrganizationTemplatePosition;
}